create library for readablity
